Convert traits to methods of Ndarr
