+++ src/ops.rs
@@ -1,4 +1,4 @@
-use crate::{scalars::{Scalar, Float}, primitives::{Broadcast, DimError, Reduce, BroadcastData}, helpers::{const_max}};
+use crate::{scalars::{Scalar, Float}, helpers::{const_max}};
 
 use super::*;
 use std::ops::*;
